refactor(types): group team domain under common/types/team
