+++ packages/desktop/src/common/adapter/teamMapper.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import type { TeamAgent, TeammateRole, TeammateStatus, TTeam, WorkspaceMode } from '../types/teamTypes';
+import type { TeamAgent, TeammateRole, TeammateStatus, TTeam, WorkspaceMode } from '../types/team/teamTypes';
 
 // ── Parameter types for team API calls ─────────────────────────────────
